feat: Refactor MLXFlux component to use BaseInput and BaseTextarea for improved input handling
